Update module names to new convention.
