+++ docs/src/API/DataContainers.md
@@ -1,0 +1,12 @@
+# DataContainers
+
+```@meta
+CurrentModule = EnsembleKalmanProcesses.DataContainers
+```
+
+```@docs
+DataContainer
+PairedDataContainer
+size
+```
+
