--- docs/src/API/DataStorage.md
@@ -1,12 +1,0 @@
-# DataStorage
-
-```@meta
-CurrentModule = EnsembleKalmanProcesses.DataStorage
-```
-
-```@docs
-DataContainer
-PairedDataContainer
-size
-```
-
