--- docs/src/API/EnsembleKalmanProcessModule.md
@@ -1,22 +1,0 @@
-# EnsembleKalmanProcessModule
-
-```@meta
-CurrentModule = EnsembleKalmanProcesses.EnsembleKalmanProcessModule
-```
-
-```@docs
-Inversion
-Sampler
-EnsembleKalmanProcess
-get_u
-get_g
-get_u_final
-get_N_iterations
-construct_initial_ensemble
-find_ekp_stepsize
-update_ensemble!
-```
-
-
-
-
