+++ docs/src/API/EnsembleKalmanProcesses.md
@@ -5,6 +5,16 @@ CurrentModule = EnsembleKalmanProcesses
 ```
 
 ```@docs
+Inversion
+Sampler
+EnsembleKalmanProcess
+get_u
+get_g
+get_u_final
+get_N_iterations
+construct_initial_ensemble
+find_ekp_stepsize
+update_ensemble!
 ```
