+++ examples/ClimateMachine/init_calibration.jl
@@ -1,9 +1,9 @@
 using Distributions
 using ArgParse
 # Import EnsembleKalmanProcesses modules
-using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
+using EnsembleKalmanProcesses
 using EnsembleKalmanProcesses.Observations
-using EnsembleKalmanProcesses.ParameterDistributionStorage
+using EnsembleKalmanProcesses.ParameterDistributions
 include(joinpath(@__DIR__, "helper_funcs.jl"))
 
 # Set parameter priors
