+++ examples/ClimateMachine/sstep_calibration.jl
@@ -4,9 +4,9 @@ using ArgParse
 using NCDatasets
 using LinearAlgebra
 # Import EnsembleKalmanProcesses modules
-using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
+using EnsembleKalmanProcesses
 using EnsembleKalmanProcesses.Observations
-using EnsembleKalmanProcesses.ParameterDistributionStorage
+using EnsembleKalmanProcesses.ParameterDistributions
 
 include(joinpath(@__DIR__, "helper_funcs.jl"))
