+++ examples/Cloudy/Cloudy_example_uki.jl
@@ -15,9 +15,9 @@ using JLD2
 using Random
 
 # Import Calibrate-Emulate-Sample modules
-using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
+using EnsembleKalmanProcesses
 using EnsembleKalmanProcesses.Observations
-using EnsembleKalmanProcesses.ParameterDistributionStorage
+using EnsembleKalmanProcesses.ParameterDistributions
 
 # Import the module that runs Cloudy
 include("DynamicalModel.jl")
@@ -123,7 +123,7 @@ for i in 1:n_samples
     y_t[:, i] = G_t .+ rand(MvNormal(μ, Γy))
 end
 
-truth = Observations.Obs(y_t, Γy, data_names)
+truth = Observations.Observation(y_t, Γy, data_names)
 truth_sample = truth.mean
 
 
@@ -159,7 +159,7 @@ for n in 1:N_iter
     println("size: ", size(ϕ_n))
     G_n = [run_dyn_model(ϕ_n[:, i], model_settings) for i in 1:size(ϕ_n)[2]]
     G_ens = hcat(G_n...)  # reformat
-    EnsembleKalmanProcessModule.update_ensemble!(ukiobj, G_ens)
+    EnsembleKalmanProcesses.update_ensemble!(ukiobj, G_ens)
     err[n] = get_error(ukiobj)[end]
     println(
         "Iteration: " *
