+++ examples/Lorenz/Lorenz_example.jl
@@ -10,10 +10,12 @@ using Random
 using JLD2
 
 # CES 
-using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
+using EnsembleKalmanProcesses
 using EnsembleKalmanProcesses.Observations
-using EnsembleKalmanProcesses.ParameterDistributionStorage
-using EnsembleKalmanProcesses.DataStorage
+using EnsembleKalmanProcesses.DataContainers
+using EnsembleKalmanProcesses.ParameterDistributions
+
+const EKP = EnsembleKalmanProcesses
 
 rng_seed = 4137
 Random.seed!(rng_seed)
@@ -195,7 +197,7 @@ end
 
 
 # Construct observation object
-truth = Observations.Obs(yt, Γy, data_names)
+truth = Observations.Observation(yt, Γy, data_names)
 truth_sample = truth.mean
 ###
 ###  Calibrate: Ensemble Kalman Inversion
@@ -214,8 +216,7 @@ N_iter = 5 # number of EKI iterations
 # initial parameters: N_params x N_ens
 initial_params = construct_initial_ensemble(priors, N_ens; rng_seed = rng_seed)
 
-ekiobj =
-    EnsembleKalmanProcessModule.EnsembleKalmanProcess(initial_params, truth_sample, truth.obs_noise_cov, Inversion())
+ekiobj = EKP.EnsembleKalmanProcess(initial_params, truth_sample, truth.obs_noise_cov, Inversion())
 
 # EKI iterations
 println("EKP inversion error:")
@@ -227,7 +228,7 @@ for i in 1:N_iter
         params_i = exp_transform(get_u_final(ekiobj))
     end
     g_ens = GModel.run_G_ensemble(params_i, lorenz_settings_G)
-    EnsembleKalmanProcessModule.update_ensemble!(ekiobj, g_ens)
+    EKP.update_ensemble!(ekiobj, g_ens)
     err[i] = get_error(ekiobj)[end] #mean((params_true - mean(params_i,dims=2)).^2)
     println("Iteration: " * string(i) * ", Error: " * string(err[i]))
 end
